Resources are now identified by a 'pointer'
A pointer is a hash containing the logical_path or id of the file, realpath on disk, and the parser for which to parse the resource. Pointers are passed around the system to uniquely identify a given file. Pointers stand in for the file's larger content body since we don't store file contents in the database, only meta-data.